Use double for storing server score to be more robust [VPNAND-1228].

+++ app/src/main/java/com/protonvpn/android/models/vpn/LoadsResponse.kt
@@ -30,7 +30,7 @@ class LoadsResponse(
 data class LoadUpdate(
     @SerialName(value = "ID") val id: String,
     @SerialName(value = "Load") val load: Float,
-    @SerialName(value = "Score") val score: Float = 0f,
+    @SerialName(value = "Score") val score: Double = 0.0,
     @SerialName(value = "Status") val status: Int = 1
 ) {
     val isOnline get() = status != 0
